Fix exception signature
I previously extended the Etcd3Exception constructor signature such
that it was no longer possible for the derived exception types to
construct with no args.  This change makes that possible again, while
still allowing an Etcd3Exception to be constructed with detail text as
well as the traditional "Bad Request" message.